+++ tapetool/helpers.py
@@ -25,4 +25,4 @@ def rms(x):
     return np.sqrt(np.sum(x**2, axis=0) / len(x))
 
 def cut(fs, t0, length):
-    return slice(fs * t0, fs * (t0 + length))
+    return slice(int(fs * t0), int(fs * (t0 + length)))
